requirements.txt
certifi==2021.5.30
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.4
MedPy==0.4.0
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264701743/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.1.48
pairwise==0.1
Pillow @ file:///tmp/build/80754af9/pillow_1625655818400/work
protobuf==3.15.6
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.3
SimpleCRF==0.1.0
SimpleITK==2.0.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboardX==2.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.59.0
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
